Include required headers in module.ac tests
Without these includes, functions exit, memset, strcmp, and strcpy are
implicitly declared, which generates warnings. On new compilers, like
the versions of clang included with Xcode 12 or later, implicitly
declared functions are treated as errors, causing the configure tests to
fail. The build later fails because the configure script made the wrong
determinations.